pyproject.toml
[project]
name = "adk-deepresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk==1.18.0",
    "langfuse>=3.10.0",
    "mypy>=1.18.2",
    "nest-asyncio>=1.6.0",
    "openinference-instrumentation-google-adk>=0.1.6",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.5",
    "streamlit>=1.51.0",
]
